requirements.txt
google-generativeai
python-dotenv
flask
pandas
openpyxl
